policy server: file reorg and settings validation
Better organize the codebase, perform settings validation.

If a policy has invalid settings, the policy server will exit with an
error.
