+++ src/main.rs
@@ -17,6 +17,9 @@ mod server;
 mod wasm_fetcher;
 mod worker;
 
+mod worker_pool;
+use worker_pool::WorkerPool;
+
 use policy_evaluator::policy::read_policies_file;
 
 mod sources;
@@ -24,8 +27,8 @@ use sources::read_sources_file;
 
 use std::fs;
 
-mod wasm;
-use crate::wasm::EvalRequest;
+mod communication;
+use communication::EvalRequest;
 
 use crate::registry::config::{DockerConfig, DockerConfigRaw};
 
@@ -210,7 +213,7 @@ fn main() {
 
     let wasm_thread = thread::spawn(move || {
         let worker_pool =
-            worker::WorkerPool::new(pool_size, policies.clone(), api_rx, barrier, boot_canary);
+            WorkerPool::new(pool_size, policies.clone(), api_rx, barrier, boot_canary);
         worker_pool.run();
     });
     // wait for all the workers to be ready, then ensure none of them had issues
